LadderProcessor 클래스 생성 : 사다리 게임 정보(플레이어, 벽, 라인 등) 처리 위한 역할(책임) 분리
당초 게임 실행 엔진 역할을 하는 GameApp에서 사다리 게임 정보를 처리하는 작업을 하고 있었는데, 이를 LadderProcessor 클래스를 별도 생성하여 역할과 책임을 분리하고자 했습니다.
