+++ ladder/view/GameDisplay.java
@@ -18,7 +18,7 @@ public class GameDisplay {
         System.out.println("최대 사다리 높이는 몇개 인가요?");
     }
 
-    public static void showLadderInfo(ArrayList<String> players, ArrayList<ArrayList<String>> ladderInfo) {
+    public static void showLadderGameInfo(ArrayList<String> players, ArrayList<ArrayList<String>> ladderInfo) {
         System.out.printf("%n  ");
         for (String player : players) {
             System.out.printf(getPaddingString(player) + " ");
@@ -31,9 +31,9 @@ public class GameDisplay {
         }
     }
 
+    // Middle Padding
     private static String getPaddingString(String player) {
         if (paddingValue <= player.length()) return player;
-        // Middle Padding
         StringBuilder sb = new StringBuilder(paddingValue);
         for (int i = 0; i < (paddingValue - player.length()) / 2; i++) {
             sb.append(" ");
@@ -44,5 +44,4 @@ public class GameDisplay {
         }
         return sb.toString();
     }
-
 }
